% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Creating aa based on sequence of codons.}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{list of characters with codons.}
}
\value{
A string of one letter abbreviation of amino acids
}
\description{
\code{mk_codons} takes a list of characters containing codons and translate it into the one latter abbreviation of the amino acid. 
A stopcodon translate to "*"
}
\examples{
dna_codons_to_aa(c("TTT", "TAA"))

}
