% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Creating codons based on DNA sequence.}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{dna}{string of nucleotides of at least length 3.}

\item{s}{positive integer between 1 and lenght of dna, dictates the stating 
position of the subsetting. Default is s = 1.}
}
\value{
list of characters with codons. If length of dna is not dividable by
3 the remaining nucleotides are omitted.
}
\description{
\code{mk_codons} takes a DNA sequence and generates a vector of codons of 3 
nucleotides. By default subseting starts at position 1, but can be changed by 
providing a value to s (needs to be a positive integer)
}
\examples{
mk_codons(dna, s = 1)

}
